# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Suskii Travels API",
    "version": "1.0.0",
    "description": "One API for the Suskii website, mobile app and admin console. Browser clients authenticate with httpOnly cookies plus an X-CSRF-Token header; mobile clients send a bearer access token. Errors use RFC 9457 problem details."
  },
  "paths": {
...
    "/v1/admin/users/{id}": {
      "get": {
        "operationId": "adminGetUser",
        "summary": "Get a user",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUser"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem400"
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          },
          "429": {
            "$ref": "#/components/responses/Problem429"
          },
          "500": {
            "$ref": "#/components/responses/Problem500"
          }
        }
      }
    },
    "/v1/admin/users/{id}/roles": {
      "put": {
        "operationId": "adminSetUserRoles",
        "summary": "Replace a user's roles",
        "description": "Audited. Signs the user out everywhere so the new roles apply immediately.",
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "description": "Required when the request is authenticated with session cookies: echo the csrf cookie value.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRolesRequestInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUser"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem400"
          },
          "401": {
            "$ref": "#/components/responses/Problem401"
          },
          "403": {
            "$ref": "#/components/responses/Problem403"
          },
          "404": {
            "$ref": "#/components/responses/Problem404"
          },
          "409": {
            "$ref": "#/components/responses/Problem409"
          },
          "429": {
            "$ref": "#/components/responses/Problem429"
          },
          "500": {
            "$ref": "#/components/responses/Problem500"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ]
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "customer",
                "super_admin",
                "operations",
                "finance",
                "support",
                "content_manager",
                "visa_officer"
              ]
            }
          },
          "mfaEnabled": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "phone",
          "displayName",
          "status",
          "roles",
          "mfaEnabled",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 9457 problem details (application/problem+json).",
        "properties": {
          "type": {
            "type": "string",
            "description": "Problem type URN, e.g. urn:suskii:problem:validation-failed"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requestId": {
            "type": "string",
            "description": "Quote this when contacting support."
          },
          "errors": {
            "type": "array",
            "description": "Field-level validation issues (validation-failed only).",
            "items": {
              "type": "object",
              "properties": {
                "location": {
                  "type": "string",
                  "enum": [
                    "body",
                    "query",
                    "params"
                  ]
                },
                "path": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "location",
                "path",
                "code",
                "message"
              ]
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
...
      "SetRolesRequestInput": {
        "type": "object",
        "properties": {
          "roles": {
            "minItems": 1,
            "maxItems": 7,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "customer",
                "super_admin",
                "operations",
                "finance",
                "support",
                "content_manager",
                "visa_officer"
              ]
            }
          }
        },
        "required": [
          "roles"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-suskii_at"
      }
    },
    "responses": {
      "Problem400": {
        "description": "Validation failed",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Problem401": {
        "description": "Authentication required",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Problem403": {
        "description": "Forbidden",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Problem404": {
        "description": "Not found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Problem409": {
        "description": "Conflict",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
...
      "Problem429": {
        "description": "Too many requests",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "Problem500": {
        "description": "Internal server error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ]
}